requirements.txt
# Flask==2.0.2
# mysql-connector-python==8.0.30
# python-dotenv==0.19.2
# Werkzeug==2.0.3   # Downgrade to a compatible version
# pytest==7.2.0 
# requests==2.26.0
# Flask-Migrate==4.0.0

Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.0
mysql-connector-python==8.0.30
python-dotenv==0.19.2
Werkzeug==2.0.3
pytest==7.2.0
requests==2.26.0
PyMySQL==1.0.3  # Required for SQLAlchemy with MySQL
